embed_fonts {extrafont}	R Documentation

Embeds fonts that are listed in the local Fontmap

Description

Embeds fonts that are listed in the local Fontmap

Usage

embed_fonts(file, format, outfile = file, options = "")

Arguments

file
Name of input file.

format
File format. (see ?embedFonts)

outfile
Name of the output file (with fonts embedded). (Default is same as input file)

options
Other arguments passed to embedFonts.

See Also

loadfonts

Examples

## Not run: 
loadfonts()
pdf('fonttest.pdf')
library(ggplot2)

p <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point()

# Run only the code below that is appropriate for your system
# On Mac and Windows, Impact should be available
p + opts(axis.title.x=theme_text(size=16, family="Impact", colour="red"))

# On Linux, Purisa may be available
p + opts(axis.title.x=theme_text(size=16, family="Purisa", colour="red"))
dev.off()

embed_fonts('fonttest.pdf', outfile='fonttest-embed.pdf')

## End(Not run)

[Package extrafont version 0.20 Index]