extrafont-package {extrafont}	R Documentation

extrafont: Tools for Using Fonts

Description

Tools to using fonts other than the standard PostScript fonts. This package makes it easy to use system TrueType fonts and with PDF or PostScript output files, and with bitmap output files in Windows. extrafont can also be used with fonts packaged specifically to be used with, such as the fontcm package, which has Computer Modern PostScript fonts with math symbols.

Author(s)

This package has been written by Winston Chang (github.com/wch/extrafont)

Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>

See Also

Useful links:

https://github.com/fbertran/extrafont

Report bugs at https://github.com/fbertran/extrafont/issues

[Package extrafont version 0.20 Index]