font_install {extrafont}	R Documentation

Install a font package and register it in the fonts database

Description

If the font package specified by fontpkg is not already installed, it will be downloaded from CRAN. The font package will then be registered in the fonts database.

Usage

font_install(fontpkg = NULL, prompt = TRUE)

Arguments

fontpkg
The name of an R package containing a font, e.g., "fontcm".

prompt
Ask user to install font from CRAN if not already installed

Examples

## Not run: 
font_install('fontcm')

## End(Not run)

[Package extrafont version 0.20 Index]