fonts {extrafont}	R Documentation

Show the fonts that are registered in the font table (and available for embedding)

Description

Show the fonts that are registered in the font table (and available for embedding)

Usage

fonts()

[Package extrafont version 0.20 Index]