loadfonts {extrafont}	R Documentation

Reads the fonttable database and registers those fonts with R

Description

This registers fonts so that they can be used with the pdf, postscript, or Windows bitmap output device. It must be run once in each R session.

Usage

loadfonts(device = c("all", "pdf", "postscript", "win"), quiet = FALSE)

Arguments

device
The output device. If "all", then it will load "pdf", "postscript", and "win" (if on Windows).

quiet
If FALSE, print a status message as each font is registered. If TRUE, don't print.

Value

A named list with up to three elements, one for each device for which fonts were loaded. Each device element is a named list, with an element for each family that was the function attempted to register with that device. The value is NULL if the function did not register the font family due to problems or because the font family was already registered. If value is the return value of windowsFonts for "win", postscriptFonts for "postscript", and pdfFonts for "pdf".

See Also

embed_fonts, #ifdef windows windowsFont, windowsFonts, #endif postscriptFonts, pdfFonts, Type1Font.

[Package extrafont version 0.20 Index]